requirements.txt
streamlit
Pillow
python-dotenv
requests
google-generativeai
langchain
langchain-community
langchain-google-genai
chromadb
PyPDF2
